% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdist_r.R
\name{cdist_r}
\alias{cdist_r}
\title{Calculate distance between two arrays}
\usage{
cdist_r(x1, x2)
}
\arguments{
\item{x1}{An array, each row is a data point.}

\item{x2}{Another array, each row is a data point.}
}
\value{
A distance matrix.
}
\description{
Calculate distance between two arrays
}
\examples{
cdist_r(matrix(c(1,1)), matrix(c(1,2)))
}
