% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_filter.R
\name{coord_filter}
\alias{coord_filter}
\title{Function to keep the points that are within to the boarder of a given coordinates}
\usage{
coord_filter(coord_mesh, coords, close_dis = 0.2, distance_to_boundary = 0)
}
\arguments{
\item{coord_mesh}{a matrix of new coordinates (e.g. super-resolution grid)}

\item{coords}{a matrix of original coordinates}

\item{close_dis}{the distance to the boundary for a point to be considered as a boundary point}

\item{distance_to_boundary}{add some extra distance to the boundary (default is 0)}
}
\value{
a logical vector indicating whether the point is kept
}
\description{
Function to keep the points that are within to the boarder of a given coordinates
}
