% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finite_difference.R
\name{finite_difference}
\alias{finite_difference}
\title{Calculate gradient using finite difference}
\usage{
finite_difference(
  coord,
  h,
  model,
  threads = 8,
  path = "./",
  prefix = "ori",
  save_file = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{coord}{coordinates of the data points, each row is a data point, sample * 2 matrix}

\item{h}{step size used in finite difference, default is 0.8* minimum distance between data points}

\item{model}{a fitted NNGP model}

\item{threads}{number of threads used in NNGP model}

\item{path}{path to save the results}

\item{prefix}{prefix of the file name}

\item{save_file}{whether to save the results}

\item{verbose}{whether to print out the progress}
}
\value{
a data.table containing the mean prediction and gradient
}
\description{
Calculate gradient using finite difference
}
