% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NNGP.R
\name{fit_NNGP}
\alias{fit_NNGP}
\title{Fit NNGP model}
\usage{
fit_NNGP(
  coords,
  y,
  neighbor = 10,
  threads = 1,
  n.samples = 500,
  verbose = FALSE,
  cov.model = "matern",
  starting = list(phi = 6, sigma.sq = 5, tau.sq = 1, nu = 2.5),
  tuning = list(phi = 0.5, sigma.sq = 0.5, tau.sq = 0.5, nu = 0.5),
  priors = list(phi.Unif = c(0.01, 300), sigma.sq.IG = c(2, 5), tau.sq.IG = c(2, 1),
    nu.Unif = c(2.00001, 5)),
  n.report = 100
)
}
\arguments{
\item{coords}{coordinates of the data points, each row is a data point, sample * 2 matrix}

\item{y}{response variable (e.g. gene expression, cell type proportion), sample * 1 matrix}

\item{neighbor}{number of neighbors used in NNGP model}

\item{threads}{number of threads used in NNGP model}

\item{n.samples}{number of samples used in posterior sampling}

\item{verbose}{whether to print out the progress}

\item{cov.model}{covariance model used in NNGP model, default is "matern"}

\item{starting}{starting values for the parameters used in NNGP model}

\item{tuning}{tuning parameters for the parameters used in NNGP model}

\item{priors}{prior distributions for the parameters used in NNGP model}

\item{n.report}{report at every n.report iterations}
}
\value{
a fitted NNGP model
}
\description{
Fit NNGP model
}
