% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_principle_curve.R
\name{format_principle_curve}
\alias{format_principle_curve}
\title{Reformat the principle curve output}
\usage{
format_principle_curve(boundary)
}
\arguments{
\item{boundary}{boundary coordinate n*4 matrix, output from elpigraph python package}
}
\value{
formatted boundary coordinate n*2 matrix
}
\description{
Reformat the principle curve output
}
