% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_bound.R
\name{reformat_bound}
\alias{reformat_bound}
\title{Format the boundary matrix to a matrix of segments}
\usage{
reformat_bound(bound)
}
\arguments{
\item{bound}{coordinate matrix of the boundary, each row is a data point (x,y), # * 2 matrix}
}
\value{
a matrix of the boundary, each row is a segment of the boundary, # * 4 matrix
}
\description{
Format the boundary matrix to a matrix of segments
}
