% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wombling.R
\name{wombling}
\alias{wombling}
\title{Wombling method to calculate the gradient of the boundary}
\usage{
wombling(
  model,
  boundary,
  h,
  threads,
  path = "./",
  prefix = "boundary0",
  save_file = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{model}{a fitted NNGP model}

\item{boundary}{boundary(a curve), each row is a point on the curve, each segment is defined by two neighboring points}

\item{h}{step size used in finite difference, default is 0.8* minimum distance between data points}

\item{threads}{number of threads used in NNGP model}

\item{path}{path to save the results}

\item{prefix}{prefix of the file name}

\item{save_file}{whether to save the results}

\item{verbose}{whether to print out the progress}
}
\value{
a data.table containing the mean prediction and gradient
}
\description{
Wombling method to calculate the gradient of the boundary
}
