% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wombling_direction.R
\name{wombling_direction}
\alias{wombling_direction}
\title{For checking the direction of wombling}
\usage{
wombling_direction(boundary)
}
\arguments{
\item{boundary}{boundary coordinate n*2 matrix}
}
\value{
a data.table containing the direction of the boundary
}
\description{
For checking the direction of wombling
}
